refactor: Rename usecases
