+++ consumatio/external/api.py
@@ -1,20 +1,20 @@
-from consumatio.usecases.number_of_watched_episodes import *
 from ariadne import ObjectType, QueryType, make_executable_schema, graphql_sync, load_schema_from_path, UnionType, MutationType
 from consumatio.external.tmdb import Tmdb
-from consumatio.usecases.movie_details import *
-from consumatio.usecases.tv_details import *
-from consumatio.usecases.season_details import *
-from consumatio.usecases.episode_details import *
-from consumatio.usecases.search_details import *
-from consumatio.usecases.popular_details import *
-from consumatio.usecases.tv_season_details import *
-from consumatio.usecases.tv_episode_details import *
-from consumatio.usecases.watch_count import *
-from consumatio.usecases.list import *
-from consumatio.usecases.watch_time import *
-from consumatio.usecases.favorite import *
-from consumatio.usecases.rating import *
-from consumatio.usecases.watch_status import *
+from consumatio.usecases.set_number_of_watched_episodes import *
+from consumatio.usecases.get_movie_details import *
+from consumatio.usecases.get_tv_details import *
+from consumatio.usecases.get_season_details import *
+from consumatio.usecases.get_episode_details import *
+from consumatio.usecases.get_search_details import *
+from consumatio.usecases.get_popular_details import *
+from consumatio.usecases.get_tv_season_details import *
+from consumatio.usecases.get_tv_episode_details import *
+from consumatio.usecases.get_watch_count import *
+from consumatio.usecases.get_list import *
+from consumatio.usecases.get_watch_time import *
+from consumatio.usecases.set_favorite import *
+from consumatio.usecases.set_rating import *
+from consumatio.usecases.set_watch_status import *
 from consumatio.external.db import Database
 from flask import Flask, request, jsonify
 from flask_cors import CORS
