+++ consumatio/usecases/get_popular_details.py
