+++ consumatio/usecases/get_search_details.py
