+++ consumatio/usecases/get_season_details.py
