+++ consumatio/usecases/get_tv_details.py
