+++ consumatio/usecases/set_favorite.py
